Make separate subaddons for arsenal configs
