+++ addons/csw/config.cpp
@@ -20,4 +20,3 @@ class CfgPatches {
 #include "CfgWeapons.hpp"
 #include "CfgMagazines.hpp"
 #include "CfgMagazineGroups.hpp"
-#include "ACE_Arsenal_Stats.hpp"
