+++ addons/explosives/compat_arsenal/ACE_Arsenal_Stats.hpp
