+++ addons/explosives/config.cpp
@@ -23,7 +23,6 @@ class CfgPatches {
 #include "CfgVehicles.hpp"
 
 #include "ACE_Triggers.hpp"
-#include "ACE_Arsenal_Stats.hpp"
 #include "ACE_Medical_Injuries.hpp"
 
 // UI stuff
