+++ addons/flashlights/config.cpp
@@ -27,4 +27,3 @@ class CfgPatches {
 #include "CfgEventHandlers.hpp"
 #include "CfgVehicles.hpp"
 #include "CfgWeapons.hpp"
-#include "ACE_Arsenal_Stats.hpp"
