+++ addons/gforces/config.cpp
@@ -21,4 +21,3 @@ class CfgPatches {
 #include "CfgEventHandlers.hpp"
 #include "CfgWeapons.hpp"
 #include "CfgVehicles.hpp"
-#include "ACE_Arsenal_Stats.hpp"
