+++ addons/logistics_wirecutter/config.cpp
@@ -17,4 +17,3 @@ class CfgPatches {
 #include "CfgEventHandlers.hpp"
 #include "CfgWeapons.hpp"
 #include "CfgVehicles.hpp"
-#include "ACE_Arsenal_Stats.hpp"
