+++ addons/reload/compat_arsenal/script_component.hpp
@@ -1,0 +1,3 @@
+#define SUBCOMPONENT arsenal
+#define SUBCOMPONENT_BEAUTIFIED Arsenal
+#include "..\script_component.hpp"
