+++ addons/trenches/config.cpp
@@ -14,7 +14,6 @@ class CfgPatches {
     };
 };
 
-#include "ACE_Arsenal_Stats.hpp"
 #include "CfgEventHandlers.hpp"
 #include "CfgVehicles.hpp"
 #include "CfgWeapons.hpp"
